shuffle code around + simplify access
