+++ src/helpers/index.ts
@@ -1,8 +1,11 @@
-export { constructFetcher as constructAxiosFetcher } from './axios';
-export { constructFetcher as constructFetchFetcher } from './fetch';
-export { constructContractCaller as constructEthersContractCaller } from './ethers';
+export { constructFetcher as constructAxiosFetcher } from './fetchers/axios';
+export { constructFetcher as constructFetchFetcher } from './fetchers/fetch';
+export {
+  constructContractCaller as constructEthersContractCaller,
+  EthersProviderDeps,
+} from './providers/ethers';
 export {
   constructContractCaller as constructWeb3ContractCaller,
-  UnpromiEvent as Web3UnpromiEvent,
-} from './web3';
+  Web3UnpromiEvent,
+} from './providers/web3';
 export { isFetcherError } from './misc';
