+++ src/helpers/token.ts
@@ -1,6 +1,4 @@
 import type { MarkOptional } from 'ts-essentials';
-import { API_URL } from './constants';
-import type { ConstructFetchInput, TokensApiResponse } from './types';
 
 /**
  * @type hex token or account address
@@ -74,31 +72,4 @@ export const constructToken = (tokenProps: ConstructTokenInput): Token => {
     network,
     ...rest,
   };
-};
-
-type GetTokens = (signal?: AbortSignal) => Promise<Token[]>;
-
-export type GetTokensFunctions = {
-  getTokens: GetTokens;
-};
-
-export const constructGetTokens = ({
-  apiURL = API_URL,
-  network,
-  fetcher,
-}: ConstructFetchInput): GetTokensFunctions => {
-  const fetchURL = `${apiURL}/tokens/${network}`;
-
-  const getTokens: GetTokens = async (signal) => {
-    const data = await fetcher<TokensApiResponse>({
-      url: fetchURL,
-      method: 'GET',
-      signal,
-    });
-
-    const tokens = data.tokens.map(constructToken);
-    return tokens;
-  };
-
-  return { getTokens };
 };
